@@ -4,8 +4,6 @@ import (
 	"database/sql"
 	"fmt"
 	"strings"
-
-	"github.com/theggv/kf2-stats-backend/pkg/common/util"
 )
 
 type MapAnalyticsService struct {
@@ -22,13 +20,13 @@ func NewMapAnalyticsService(db *sql.DB) *MapAnalyticsService {
 
 func (s *MapAnalyticsService) GetMapAnalytics(
 	req MapAnalyticsRequest,
-) (*[]MapAnalytics, error) {
+) ([]*MapAnalytics, error) {
 	limit := clampLimit(req.Limit)
 
 	conds := make([]string, 0)
-	args := make([]interface{}, 0)
+	args := make([]any, 0)
 
-	conds = append(conds, "session.is_completed = TRUE")
+	conds = append(conds, "session.started_at is not null")
 
 	if req.ServerId != 0 {
 		conds = append(conds, "session.server_id = ?")
@@ -38,30 +36,52 @@ func (s *MapAnalyticsService) GetMapAnalytics(
 	conds = append(conds, "DATE(session.started_at) BETWEEN ? AND ?")
 	args = append(args, req.From.Format("2006-01-02"), req.To.Format("2006-01-02"))
 
-	sql := fmt.Sprintf(`
-		SELECT
-			maps.id,
-			maps.name,
-			count(session.id) AS times_played
-		FROM session
-		INNER JOIN maps ON maps.id = session.map_id
-		WHERE %v
-		GROUP BY session.map_id
-		ORDER BY times_played DESC
-		LIMIT %v`, strings.Join(conds, " AND "), limit,
+	stmt := fmt.Sprintf(`
+		WITH maps_rating AS (
+			SELECT
+				maps.id as map_id,
+				maps.name as map_name,
+				count(session.id) as value
+			FROM session
+			INNER JOIN maps ON maps.id = session.map_id
+			WHERE %v
+			GROUP BY session.map_id
+			ORDER BY value DESC
+			LIMIT %v
+		), other AS (
+			SELECT count(*) as value
+			FROM session
+			LEFT JOIN rating ON rating.map_id = session.map_id
+			WHERE rating.map_id is null AND %v
+		), result AS (
+			SELECT
+				maps.id as map_id,
+				maps.name as map_name,
+				cte.value as value
+			FROM rating cte
+			INNER JOIN maps ON maps.id = cte.map_id
+
+			UNION ALL
+
+			SELECT
+				0 as map_id,
+				"Other" as map_name,
+				value
+			FROM other
+			WHERE value > 0
+		)
+		SELECT map_id, map_name, value 
+		FROM result
+		`,
+		strings.Join(conds, " AND "), limit, strings.Join(conds, " AND "),
 	)
 
-	stmt, err := s.db.Prepare(sql)
+	rows, err := s.db.Query(stmt, args...)
 	if err != nil {
 		return nil, err
 	}
 
-	rows, err := stmt.Query(args...)
-	if err != nil {
-		return nil, err
-	}
-
-	items := []MapAnalytics{}
+	items := []*MapAnalytics{}
 	for rows.Next() {
 		item := MapAnalytics{}
 
@@ -70,52 +90,10 @@ func (s *MapAnalyticsService) GetMapAnalytics(
 			return nil, err
 		}
 
-		items = append(items, item)
+		items = append(items, &item)
 	}
 
-	mapIds := []int{}
-	for _, item := range items {
-		mapIds = append(mapIds, item.MapId)
-	}
-
-	if len(mapIds) > 0 {
-		sql := fmt.Sprintf(`
-			SELECT count(session.id) AS times_played
-			FROM session
-			INNER JOIN maps ON maps.id = session.map_id
-			WHERE %v AND session.map_id not in (%v)`,
-			strings.Join(conds, " AND "),
-			util.IntArrayToString(mapIds, ","),
-		)
-
-		stmt, err := s.db.Prepare(sql)
-		if err != nil {
-			return nil, err
-		}
-
-		rows, err := stmt.Query(args...)
-		if err != nil {
-			return nil, err
-		}
-
-		for rows.Next() {
-			other := MapAnalytics{
-				MapId:   0,
-				MapName: "Other",
-			}
-
-			err = rows.Scan(&other.Count)
-			if err != nil {
-				return nil, err
-			}
-
-			if other.Count > 0 {
-				items = append(items, other)
-			}
-		}
-	}
-
-	return &items, nil
+	return items, nil
 }
 
 func clampLimit(limit int) int {
