@@ -1,7 +1,6 @@
 package maps
 
 import (
-	"fmt"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
@@ -22,18 +21,16 @@ func (c *controller) getMapAnalytics(ctx *gin.Context) {
 	var req MapAnalyticsRequest
 	if err := ctx.ShouldBindBodyWith(&req, binding.JSON); err != nil {
 		ctx.String(http.StatusBadRequest, err.Error())
-		fmt.Printf("%v\n", err.Error())
 		return
 	}
 
 	items, err := c.service.GetMapAnalytics(req)
 	if err != nil {
 		ctx.String(http.StatusBadRequest, err.Error())
-		fmt.Printf("%v\n", err.Error())
 		return
 	}
 
 	ctx.JSON(http.StatusOK, &MapAnalyticsResponse{
-		Items: *items,
+		Items: items,
 	})
 }
