@@ -22,7 +22,6 @@ func (c *controller) getSessionCount(ctx *gin.Context) {
 	var req SessionCountRequest
 	if err := ctx.ShouldBindBodyWith(&req, binding.JSON); err != nil {
 		ctx.String(http.StatusBadRequest, err.Error())
-		fmt.Printf("%v\n", err.Error())
 		return
 	}
 
@@ -34,7 +33,7 @@ func (c *controller) getSessionCount(ctx *gin.Context) {
 	}
 
 	ctx.JSON(http.StatusOK, &SessionCountResponse{
-		Items: *items,
+		Items: items,
 	})
 }
 
@@ -48,19 +47,17 @@ func (c *controller) getUsageInMinutes(ctx *gin.Context) {
 	var req UsageInMinutesRequest
 	if err := ctx.ShouldBindBodyWith(&req, binding.JSON); err != nil {
 		ctx.String(http.StatusBadRequest, err.Error())
-		fmt.Printf("%v\n", err.Error())
 		return
 	}
 
 	items, err := c.service.GetUsageInMinutes(req)
 	if err != nil {
 		ctx.String(http.StatusBadRequest, err.Error())
-		fmt.Printf("%v\n", err.Error())
 		return
 	}
 
 	ctx.JSON(http.StatusOK, &UsageInMinutesResponse{
-		Items: *items,
+		Items: items,
 	})
 }
 
@@ -74,19 +71,17 @@ func (c *controller) getPlayersOnline(ctx *gin.Context) {
 	var req PlayersOnlineRequest
 	if err := ctx.ShouldBindBodyWith(&req, binding.JSON); err != nil {
 		ctx.String(http.StatusBadRequest, err.Error())
-		fmt.Printf("%v\n", err.Error())
 		return
 	}
 
 	items, err := c.service.GetPlayersOnline(req)
 	if err != nil {
 		ctx.String(http.StatusBadRequest, err.Error())
-		fmt.Printf("%v\n", err.Error())
 		return
 	}
 
 	ctx.JSON(http.StatusOK, &PlayersOnlineResponse{
-		Items: *items,
+		Items: items,
 	})
 }
 
@@ -99,7 +94,6 @@ func (c *controller) getPopularServers(ctx *gin.Context) {
 	res, err := c.service.GetPopularServers()
 	if err != nil {
 		ctx.String(http.StatusBadRequest, err.Error())
-		fmt.Printf("%v\n", err.Error())
 		return
 	}
 
@@ -115,7 +109,6 @@ func (c *controller) getCurrentOnline(ctx *gin.Context) {
 	res, err := c.service.GetCurrentOnline()
 	if err != nil {
 		ctx.String(http.StatusBadRequest, err.Error())
-		fmt.Printf("%v\n", err.Error())
 		return
 	}
 
