@@ -3,6 +3,7 @@ package users
 import (
 	"github.com/chenyahui/gin-cache/persist"
 	"github.com/gin-gonic/gin"
+	"github.com/theggv/kf2-stats-backend/pkg/common/middleware"
 )
 
 func RegisterRoutes(
@@ -18,8 +19,17 @@ func RegisterRoutes(
 
 	routes.POST("/users", controller.getUserAnalytics)
 	routes.POST("/users/perks", controller.getPerksAnalytics)
-	routes.POST("/users/perks/playtime", controller.getPlaytimeHist)
-	routes.POST("/users/perks/accuracy", controller.getAccuracyHist)
-	routes.POST("/users/teammates", controller.getTeammates)
-
+	routes.POST("/users/perks/playtime",
+		middleware.OptionalAuthMiddleWave, controller.getPlaytimeHist)
+	routes.POST("/users/perks/accuracy",
+		middleware.OptionalAuthMiddleWave, controller.getAccuracyHist)
+	routes.POST("/users/teammates",
+		middleware.OptionalAuthMiddleWave, controller.getTeammates)
+	routes.POST("/users/maps", controller.getPlayedMaps)
+	routes.POST("/users/sessions",
+		middleware.OptionalAuthMiddleWave, controller.getUserSessions)
+	routes.POST("/users/lastseen",
+		middleware.AuthMiddleWave, controller.getLastSeenUsers)
+	routes.POST("/users/lastgameswithuser",
+		middleware.AuthMiddleWave, controller.getLastGamesWithUser)
 }
