@@ -7,27 +7,33 @@ import (
 	_ "github.com/go-sql-driver/mysql"
 )
 
-func NewDBInstance(user, pass, host, db string, port int) *sql.DB {
+type DBConnection struct {
+	Conn *sql.DB
+}
+
+func NewDBInstance(user, pass, host, db string, port int) (*DBConnection, error) {
 	connString := fmt.Sprintf("%s:%s@tcp(%s:%d)/%s?charset=utf8&parseTime=True&loc=Local&multiStatements=True",
 		user, pass, host, port, db,
 	)
 
 	instance, err := sql.Open("mysql", connString)
 	if err != nil {
+		return nil, err
+	}
+
+	return &DBConnection{Conn: instance}, nil
+}
+
+func (c *DBConnection) InitTables() {
+	if err := initSchema(c.Conn); err != nil {
 		panic(err)
 	}
 
-	if err := initSchema(instance); err != nil {
+	if err := initTriggers(c.Conn); err != nil {
 		panic(err)
 	}
 
-	if err := initTriggers(instance); err != nil {
+	if err := initStored(c.Conn); err != nil {
 		panic(err)
 	}
-
-	if err := initStored(instance); err != nil {
-		panic(err)
-	}
-
-	return instance
 }
