@@ -22,7 +22,7 @@ func initStored(db *sql.DB) error {
 			) as tbl on session.server_id = tbl.server_id
 			SET status = -1
 			WHERE 
-				session.id <> 0 AND session.id NOT IN (tbl.max_id) AND 
+				session.id NOT IN (tbl.max_id) AND 
 				status IN (0, 1);
 		END;
 	`)
@@ -35,8 +35,8 @@ func initStored(db *sql.DB) error {
 			INNER JOIN session_game_data gd ON gd.session_id = session.id
 			SET session.status = -1
 			WHERE 
-				session.id <> 0 AND 
 				session.status IN (0, 1) AND 
+				session.started_at >= CURRENT_TIMESTAMP - INTERVAL 30 DAY AND
 				timestampdiff(MINUTE, gd.updated_at, CURRENT_TIMESTAMP) > minutes;
 		END;
 	`)
@@ -44,17 +44,16 @@ func initStored(db *sql.DB) error {
 		DROP PROCEDURE IF EXISTS delete_empty_sessions;
 		CREATE PROCEDURE delete_empty_sessions()
 		BEGIN
-			DELETE FROM session WHERE id IN (
-				SELECT id FROM (
-					SELECT distinct session.id, session.status
-					FROM session
-						LEFT JOIN wave_stats ws ON ws.session_id = session.id
-						LEFT JOIN wave_stats_player wsp ON wsp.stats_id = ws.id
-					WHERE session.status IN (-1, 2, 3)
-					GROUP BY session.id
-					HAVING count(wsp.id) = 0
-				) t
-			);
+			DELETE FROM session 
+			WHERE
+				session.status IN (-1, 2, 3) AND
+				session.started_at >= CURRENT_TIMESTAMP - INTERVAL 30 DAY AND
+				NOT EXISTS (
+					SELECT 1
+					FROM wave_stats ws
+					INNER JOIN wave_stats_player wsp ON wsp.stats_id = ws.id
+					WHERE ws.session_id = session.id
+				);
 		END;
 	`)
 	tx.Exec(`
@@ -225,7 +224,6 @@ func initStored(db *sql.DB) error {
 		DROP PROCEDURE IF EXISTS insert_session_aggregated;
 		CREATE PROCEDURE insert_session_aggregated(session_id INT)
 		BEGIN
-			DO SLEEP(3);
 			INSERT IGNORE INTO session_aggregated (
 				session_id, user_id, perk, 
 				playtime_seconds, waves_played, deaths, 
@@ -284,6 +282,183 @@ func initStored(db *sql.DB) error {
 					GROUP BY session.id, wsp.player_id, wsp.perk
 				) t ON aggr.session_id = t.session_id AND aggr.user_id = t.user_id AND aggr.perk = t.perk
 			);
+
+			CALL update_user_stats_weekly(session_id);
+		END;
+	`)
+	tx.Exec(`
+		DROP PROCEDURE IF EXISTS update_user_stats_weekly;
+		CREATE PROCEDURE update_user_stats_weekly(session_id INT)
+		BEGIN
+			INSERT IGNORE INTO user_weekly_stats_total SELECT * FROM (
+				SELECT
+					YEARWEEK(session.started_at) as period, 
+
+					session.server_id as server_id,
+					wsp.player_id as user_id, 
+
+					1 as total_games,
+					count(*) as total_waves,
+					sum(timestampdiff(SECOND, ws.started_at, ws.completed_at)) as playtime_seconds,
+					sum(is_dead = 1) as deaths, 
+
+					sum(shots_fired) as shots_fired, 
+					sum(shots_hit) as shots_hit, 
+					sum(shots_hs) as shots_hs, 
+
+					sum(dosh_earned) as dosh_earned, 
+					sum(heals_given) as heals_given, 
+					sum(heals_recv) as heals_recv, 
+					sum(damage_dealt) as damage_dealt, 
+					sum(damage_taken) as damage_taken, 
+
+					sum(k.scrake) + sum(k.fp) + sum(k.qp) as large_kills, 
+					sum(k.cyst) + sum(k.alpha_clot) + sum(k.slasher) + 
+					sum(k.stalker) + sum(k.crawler) + sum(k.gorefast) + 
+					sum(k.rioter) + sum(k.elite_crawler) + sum(k.gorefiend) + 
+					sum(k.siren) + sum(k.bloat) + sum(k.edar) + sum(k.husk_n) + sum(k.husk_b) + 
+					sum(k.scrake) + sum(k.fp) + sum(k.qp) + sum(k.boss) + sum(k.custom) as total_kills,
+
+					session.id as max_damage_session_id,
+					sum(damage_dealt) as max_damage
+				FROM session
+				INNER JOIN wave_stats ws ON ws.session_id = session.id
+				INNER JOIN wave_stats_player wsp ON wsp.stats_id = ws.id
+				INNER JOIN wave_stats_player_kills k ON k.player_stats_id = wsp.id
+				WHERE session.id = session_id
+				GROUP BY session.id, wsp.player_id
+			) as new
+			ON DUPLICATE KEY UPDATE 
+				user_weekly_stats_total.total_games = user_weekly_stats_total.total_games + new.total_games,
+				user_weekly_stats_total.total_waves = user_weekly_stats_total.total_waves + new.total_waves,
+				user_weekly_stats_total.playtime_seconds = user_weekly_stats_total.playtime_seconds + new.playtime_seconds,
+				user_weekly_stats_total.deaths = user_weekly_stats_total.deaths + new.deaths,
+
+				user_weekly_stats_total.shots_fired = user_weekly_stats_total.shots_fired + new.shots_fired,
+				user_weekly_stats_total.shots_hit = user_weekly_stats_total.shots_hit + new.shots_hit,
+				user_weekly_stats_total.shots_hs = user_weekly_stats_total.shots_hs + new.shots_hs,
+
+				user_weekly_stats_total.dosh_earned = user_weekly_stats_total.dosh_earned + new.dosh_earned,
+				user_weekly_stats_total.heals_given = user_weekly_stats_total.heals_given + new.heals_given,
+				user_weekly_stats_total.heals_recv = user_weekly_stats_total.heals_recv + new.heals_recv,
+
+				user_weekly_stats_total.damage_dealt = user_weekly_stats_total.damage_dealt + new.damage_dealt,
+				user_weekly_stats_total.damage_taken = user_weekly_stats_total.damage_taken + new.damage_taken,
+				
+				user_weekly_stats_total.large_kills = user_weekly_stats_total.large_kills + new.large_kills,
+				user_weekly_stats_total.total_kills = user_weekly_stats_total.total_kills + new.total_kills,
+
+				max_damage_session_id = new.max_damage_session_id,
+				max_damage = new.max_damage;
+
+			INSERT IGNORE INTO user_weekly_stats_perk SELECT * FROM (
+				SELECT
+					YEARWEEK(session.started_at) as period, 
+
+					session.server_id as server_id,
+					wsp.player_id as user_id, 
+					wsp.perk as perk,
+
+					1 as total_games,
+					count(*) as total_waves,
+					sum(timestampdiff(SECOND, ws.started_at, ws.completed_at)) as playtime_seconds,
+					sum(is_dead = 1) as deaths, 
+
+					sum(shots_fired) as shots_fired, 
+					sum(shots_hit) as shots_hit, 
+					sum(shots_hs) as shots_hs, 
+
+					sum(dosh_earned) as dosh_earned, 
+					sum(heals_given) as heals_given, 
+					sum(heals_recv) as heals_recv, 
+					sum(damage_dealt) as damage_dealt, 
+					sum(damage_taken) as damage_taken, 
+
+					sum(zedtime_count) as zedtime_count, 
+					sum(zedtime_length) as zedtime_length,
+
+					0 as buffs_active_length, 
+					0 as buffs_total_length,
+
+					sum(k.scrake) + sum(k.fp) + sum(k.qp) as large_kills, 
+					sum(k.cyst) + sum(k.alpha_clot) + sum(k.slasher) + 
+					sum(k.stalker) + sum(k.crawler) + sum(k.gorefast) + 
+					sum(k.rioter) + sum(k.elite_crawler) + sum(k.gorefiend) + 
+					sum(k.siren) + sum(k.bloat) + sum(k.edar) + sum(k.husk_n) + sum(k.husk_b) + 
+					sum(k.scrake) + sum(k.fp) + sum(k.qp) + sum(k.boss) + sum(k.custom) as total_kills,
+
+					session.id as max_damage_session_id,
+					sum(damage_dealt) as max_damage
+				FROM session
+				INNER JOIN wave_stats ws ON ws.session_id = session.id
+				INNER JOIN wave_stats_player wsp ON wsp.stats_id = ws.id
+				INNER JOIN wave_stats_player_kills k ON k.player_stats_id = wsp.id
+				WHERE session.id = session_id
+				GROUP BY session.id, wsp.player_id, wsp.perk
+			) as new
+			ON DUPLICATE KEY UPDATE
+				user_weekly_stats_perk.total_games = user_weekly_stats_perk.total_games + new.total_games,
+				user_weekly_stats_perk.total_waves = user_weekly_stats_perk.total_waves + new.total_waves,
+				user_weekly_stats_perk.playtime_seconds = user_weekly_stats_perk.playtime_seconds + new.playtime_seconds,
+				user_weekly_stats_perk.deaths = user_weekly_stats_perk.deaths + new.deaths,
+
+				user_weekly_stats_perk.shots_fired = user_weekly_stats_perk.shots_fired + new.shots_fired,
+				user_weekly_stats_perk.shots_hit = user_weekly_stats_perk.shots_hit + new.shots_hit,
+				user_weekly_stats_perk.shots_hs = user_weekly_stats_perk.shots_hs + new.shots_hs,
+
+				user_weekly_stats_perk.dosh_earned = user_weekly_stats_perk.dosh_earned + new.dosh_earned,
+				user_weekly_stats_perk.heals_given = user_weekly_stats_perk.heals_given + new.heals_given,
+				user_weekly_stats_perk.heals_recv = user_weekly_stats_perk.heals_recv + new.heals_recv,
+
+				user_weekly_stats_perk.damage_dealt = user_weekly_stats_perk.damage_dealt + new.damage_dealt,
+				user_weekly_stats_perk.damage_taken = user_weekly_stats_perk.damage_taken + new.damage_taken,
+				
+				user_weekly_stats_perk.large_kills = user_weekly_stats_perk.large_kills + new.large_kills,
+				user_weekly_stats_perk.total_kills = user_weekly_stats_perk.total_kills + new.total_kills,
+
+				user_weekly_stats_perk.zedtime_count = user_weekly_stats_perk.zedtime_count + new.zedtime_count,
+				user_weekly_stats_perk.zedtime_length = user_weekly_stats_perk.zedtime_length + new.zedtime_length,
+
+				user_weekly_stats_perk.buffs_active_length = user_weekly_stats_perk.buffs_active_length + new.buffs_active_length,
+				user_weekly_stats_perk.buffs_total_length = user_weekly_stats_perk.buffs_total_length + new.buffs_total_length,
+
+				user_weekly_stats_perk.large_kills = user_weekly_stats_perk.large_kills + new.large_kills,
+				user_weekly_stats_perk.total_kills = user_weekly_stats_perk.total_kills + new.total_kills,
+
+				max_damage_session_id = new.max_damage_session_id,
+				max_damage = new.max_damage;
+		END;
+	`)
+	tx.Exec(`
+		DROP PROCEDURE IF EXISTS fill_weekly_user_stats;
+		CREATE PROCEDURE fill_weekly_user_stats()
+		BEGIN
+			DECLARE session_id INT DEFAULT NULL;
+			DECLARE done TINYINT DEFAULT FALSE;
+
+			DECLARE sessions_cursor 
+				CURSOR FOR SELECT id FROM session WHERE is_completed;
+				
+			DECLARE CONTINUE HANDLER FOR NOT FOUND SET done = TRUE;
+
+			START TRANSACTION;
+
+			OPEN sessions_cursor;
+
+			sessions_loop:
+			LOOP
+				FETCH NEXT FROM sessions_cursor INTO session_id;
+
+				IF done THEN
+					LEAVE sessions_loop; 
+				ELSE
+					CALL update_user_stats_weekly(session_id);
+				END IF;
+			END LOOP;
+
+			CLOSE sessions_cursor;
+
+			COMMIT;
 		END;
 	`)
 
