@@ -51,8 +51,6 @@ func NewMatchesService(db *sql.DB) *MatchesService {
 		db: db,
 	}
 
-	service.setupTasks()
-
 	return &service
 }
 
@@ -410,6 +408,10 @@ func (s *MatchesService) GetMatchAggregatedStats(sessionId int) (*GetMatchAggreg
 			&stats.Kills, &stats.LargeKills, &stats.HuskRages,
 		)
 
+		if err != nil {
+			return nil, err
+		}
+
 		players = append(players, stats)
 	}
 
@@ -426,11 +428,11 @@ func (s *MatchesService) GetMatchLiveData(sessionId int) (*GetMatchLiveDataRespo
 			wave, is_trader_time, zeds_left,
 			spawn_cycle, max_monsters, wave_size_fakes, zeds_type
 		FROM session_game_data gd
-		LEFT JOIN session_game_data_cd cd ON cd.session_id = gd.session_id
+		LEFT JOIN session_game_data_extra cd ON cd.session_id = gd.session_id
 		WHERE gd.session_id = ?`
 
 	gameData := models.GameData{}
-	cdData := models.CDGameData{}
+	cdData := models.ExtraGameData{}
 
 	err := s.db.QueryRow(sql, sessionId).Scan(
 		&gameData.MaxPlayers, &gameData.PlayersOnline, &gameData.PlayersAlive,
@@ -448,6 +450,9 @@ func (s *MatchesService) GetMatchLiveData(sessionId int) (*GetMatchLiveDataRespo
 
 	var status models.GameStatus
 	err = s.db.QueryRow(`SELECT status FROM session WHERE id = ?`, sessionId).Scan(&status)
+	if err != nil {
+		return nil, err
+	}
 
 	res := GetMatchLiveDataResponse{
 		Status:     status,
@@ -476,6 +481,9 @@ func (s *MatchesService) GetMatchLiveData(sessionId int) (*GetMatchLiveDataRespo
 		where current_session_id = ?
 		`, sessionId,
 	)
+	if err != nil {
+		return nil, err
+	}
 
 	steamIdSet := make(map[string]bool)
 	for rows.Next() {
@@ -487,6 +495,9 @@ func (s *MatchesService) GetMatchLiveData(sessionId int) (*GetMatchLiveDataRespo
 			&item.Health, &item.Armor,
 			&item.IsSpectator,
 		)
+		if err != nil {
+			return nil, err
+		}
 
 		if item.AuthType == models.Steam {
 			steamIdSet[item.AuthId] = true
