@@ -118,21 +118,31 @@ func (s *ServerService) GetRecentUsers(req RecentUsersRequest) (*RecentUsersResp
 	conds = append(conds, fmt.Sprintf("session.server_id = %v", req.ServerId))
 
 	sql := fmt.Sprintf(`
+		WITH ranking AS (
+			SELECT 
+				wsp.player_id user_id,
+				wsp.id wsp_id,
+				DENSE_RANK() OVER w as rating
+			FROM session
+			INNER JOIN wave_stats ws ON ws.session_id = session.id
+			INNER JOIN wave_stats_player wsp ON wsp.stats_id = ws.id
+			WHERE %v
+			WINDOW w AS (PARTITION BY wsp.player_id ORDER BY wsp.id DESC)
+			ORDER BY wsp.id DESC
+		)
 		SELECT 
 			users.id,
 			users.name,
 			users.auth_id,
 			users.auth_type,
-			max(session.id),
-			max(wsp.id),
-			max(wsp.created_at)
-		FROM session
-		INNER JOIN wave_stats ws ON ws.session_id = session.id
-		INNER JOIN wave_stats_player wsp ON wsp.stats_id = ws.id
-		INNER JOIN users ON wsp.player_id = users.id
-		WHERE %v
-		GROUP BY users.id
-		ORDER BY max(wsp.id) DESC
+			ws.session_id,
+			wsp.id wsp_id,
+			wsp.created_at
+		FROM ranking
+		INNER JOIN wave_stats_player wsp ON wsp.id = ranking.wsp_id
+		INNER JOIN wave_stats ws ON ws.id = wsp.stats_id
+		INNER JOIN users ON users.id = wsp.player_id
+		WHERE rating = 1
 		LIMIT %v, %v
 		`, strings.Join(conds, " AND "), page*limit, limit,
 	)
@@ -200,10 +210,9 @@ func (s *ServerService) GetRecentUsers(req RecentUsersRequest) (*RecentUsersResp
 
 	// Prepare count query
 	sql = fmt.Sprintf(`
-		SELECT count(distinct users.id) FROM session
+		SELECT count(distinct wsp.player_id) FROM session
 		INNER JOIN wave_stats ws ON ws.session_id = session.id
 		INNER JOIN wave_stats_player wsp ON wsp.stats_id = ws.id
-		INNER JOIN users ON wsp.player_id = users.id
 		WHERE %v`,
 		strings.Join(conds, " AND "),
 	)
@@ -251,7 +260,7 @@ func (s *ServerService) getSessions(
 		INNER JOIN wave_stats_player wsp ON wsp.stats_id = ws.id
 		INNER JOIN maps ON maps.id = session.map_id
 		INNER JOIN session_game_data gd ON gd.session_id = session.id
-		LEFT JOIN session_game_data_cd cd ON cd.session_id = session.id
+		LEFT JOIN session_game_data_extra cd ON cd.session_id = session.id
 		WHERE wsp.id IN (%v)
 		`, util.IntArrayToString(wspIds, ","),
 	)
@@ -265,7 +274,7 @@ func (s *ServerService) getSessions(
 	items := make(map[int]*RecentUsersResponseUserSession)
 	for rows.Next() {
 		item := RecentUsersResponseUserSession{}
-		cdData := models.CDGameData{}
+		cdData := models.ExtraGameData{}
 
 		rows.Scan(
 			&item.PlayerId, &item.Id, &item.Mode, &item.Length, &item.Difficulty, &item.Status, &item.Wave,
