@@ -41,8 +41,8 @@ type PlayerLiveData struct {
 type UpdateGameDataRequest struct {
 	SessionId int `json:"session_id"`
 
-	GameData models.GameData    `json:"game_data"`
-	CDData   *models.CDGameData `json:"cd_data"`
+	GameData models.GameData       `json:"game_data"`
+	CDData   *models.ExtraGameData `json:"cd_data"`
 
 	Players []PlayerLiveData `json:"players"`
 }
