@@ -134,9 +134,9 @@ func (s *StatsService) createWaveStatsPlayer(statsId int, req *CreateWaveStatsRe
 	return err
 }
 
-func (s *StatsService) createWaveStatsCD(statsId int, req *models.CDGameData) error {
+func (s *StatsService) createWaveStatsCD(statsId int, req *models.ExtraGameData) error {
 	_, err := s.db.Exec(`
-		INSERT INTO wave_stats_cd (
+		INSERT INTO wave_stats_extra (
 			stats_id, spawn_cycle, max_monsters, wave_size_fakes, zeds_type) 
 		VALUES (?, ?, ?, ?, ?)`,
 		statsId,
