@@ -12,8 +12,7 @@ func RegisterRoutes(r *gin.RouterGroup, service *UserService) {
 
 	routes := r.Group("/users")
 
-	routes.POST("/", middleware.AuthMiddleware, controller.create)
+	routes.POST("/", middleware.MutatorAuthMiddleWave, controller.create)
 	routes.GET("/:id/detailed", controller.getUserDetailed)
 	routes.POST("/filter", controller.filter)
-	routes.POST("/sessions", controller.getUserSessions)
 }
